[graphite] Fix the specified type does not match the argument type error
Bug: skia:14392
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/713236
Auto-Submit: Ho Cheung <uioptt24@gmail.com>
Commit-Queue: Nicolette Prevost <nicolettep@google.com>

+++ include/private/gpu/graphite/DawnTypesPriv.h
@@ -27,7 +27,9 @@ struct DawnTextureSpec {
     }
 
     SkString toString() const {
-        return SkStringPrintf("format=0x%08X,usage=0x%08X", fFormat, fUsage);
+        return SkStringPrintf("format=0x%08X,usage=0x%08X",
+                              static_cast<unsigned int>(fFormat),
+                              static_cast<unsigned int>(fUsage));
     }
 
     wgpu::TextureFormat fFormat;
